Move static methods to utils
